+++ 00 Core/scripts/PotentialCharacterProgression/info.lua
@@ -7,6 +7,6 @@ return {
     name = 'PotentialCharacterProgression',
     version = '1.0.2',
     saveVersion = 1,
-    settingsVersion = 2,
+    settingsVersion = 3,
     minApiVersion = 68
 }